pyproject.toml
[tool.poetry]
name = "brain-signals"
version = "0.1.0"
description = ""
authors = ["As17-01 <aseliverstov@nes.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
ipykernel = "^6.29.5"
numpy = "^1.24.3"
pandas = "^2.0.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.3.0"
tqdm = "^4.66.5"
loguru = "^0.7.2"
black = "^24.10.0"
tensorflow = "^2.18.0"
torch = "^2.7.1"
torcheval = "^0.0.7"
shap = "^0.46.0"
keras = "^3.8.0"
librosa = "^0.10.2.post1"
ncps = "^1.0.1"
ipyparams = "^0.2.1"
linearmodels = "^6.1"
hydra-core = "^1.3.2"
hydra-slayer = "^0.5.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
